pyproject.toml
[project]
name = "resume-analysis-tool"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.12",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "langchain-community>=0.3.26",
    "langchain-openai>=0.3.27",
    "langgraph>=0.5.0",
    "pypdf>=5.6.1",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.0.0",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "uvicorn[standard]>=0.34.3",
]

[dependency-groups]
dev = [
    "honcho>=2.0.0",
]
